feat: Open source Streammed
This commits open sources Streammed, a Streamlit-powered application
for detecting medication names in user-provided text using annotations.
The web UI allows users to provide textual input and select the model to
use for generating the annotations, with 2 models currently supported:
* Med7
* Random

The application consists of a backend, implemented with FastAPI and
exposing a single endpoint for generating annotations, and a frontend,
implemented with Streamlit to support receiving text, rendering the
annotated output highlighting medication names, adding or removing
annotations, and calculating the model precision. Usage instructions,
as well as screenshots can be found in the README.

The 2 components are packaged into separate images with Docker, while
the application can be spin up with docker-compose. A test suite is
included for both the backend and the frontend inside the 'tests'
directory and is run as part of the configured CI pipeline on GitHub,
along with the specified linting checks.

Signed-off-by: Phoevos Kalemkeris <phoevos@protonmail.com>